android: refactor video capture and move it to VideoCaptureController
This serves 2 purposes:

- reduce the complexity in GetUserMediaImpl
- pave the way for other capture sources
